Residual Blocks
After the first masked convolution the model has a series of residual blocks (The architecture picture above has only 1 residual block).

To implement a residual block:

1. Take input of shape (height, width, filters).
2. Halve the filters with a (1,1) convolution.
3. Apply a (3,3) masked convolution of type B.
4. Scale the filters back to original with (1,1) convolution.
5. Merge the original input with the convolutions.

The reason for cutting the filters by half and then scaling back to original is because it is a good way to get a computational boost while not significally reducing model performance.

We implemented a residual block in Keras.

We will want to stack those residual blocks in our model, so we created a simple layer for that.