Output, Compile, Train (Load MNIST)
Since we have just one channel, we can convolve the pixels with a convolution of size (1,1) with a single filter and then sigmoid its output. The output of the sigmoid should be a 2D array with an exact shape as the input ((28, 28, 1) for MNIST), with each point in the 2D array representing a (grayscale) color value.

I chose Nadam quite arbitrarily, you can go with any optimizer you like. Since we use sigmoid in our output activations our loss should be binary_crossentropy.

We compiled the model.

We loaded the MNIST data, change data type to float and normalize it. We ignored the labels of MNIST because they are not useful for our case, our model is not a classifier.

Note that I concatenated the training and test data, I do that to have more data to help with training, however, if you need validation data, feel free to not concatenate them.